requirements.txt
google==3.0.0
google-genai==1.20.0
python-dotenv==1.1.0
